Minecraft Autofishins script/requirements.txt
# Python version
python==3.12

# Required libraries and their versions
opencv-python==4.10.0
numpy==1.25.2
pyautogui==0.9.53
python-dotenv==1.0.0
screeninfo==0.8.0
mss==9.0.1

# Pip install (installs all the required libraries) "pip install -r requirements.txt"
